requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
python-dotenv
requests